requirements.txt
blinker==1.9.0
boto3==1.28.44
botocore==1.31.85
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.4.0
google-api-core==2.24.0
google-auth==2.38.0
google-generativeai==0.3.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.62.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
Pillow==10.0.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
s3transfer==0.6.2
six==1.17.0
SQLAlchemy==2.0.37
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.0.7
Werkzeug==3.1.3
